Misc changes

+++ src/core/file_sys/disk_archive.cpp
@@ -53,7 +53,7 @@ bool DiskFile::Close() const {
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
 DiskDirectory::DiskDirectory(const std::string& path) {
-    unsigned size = FileUtil::ScanDirectoryTree(path, directory);
+    u64 size = FileUtil::ScanDirectoryTree(path, directory);
     directory.size = size;
     directory.isDirectory = true;
     children_iterator = directory.children.begin();
